src/com/olivejua/tree/BalancedBinaryTree.java
package com.olivejua.tree;

/**
 * https://leetcode.com/problems/balanced-binary-tree
 */
public class BalancedBinaryTree {

    public boolean isBalanced(TreeNode root) {
        if (root == null) {
            return true;
        }

        int height = dfs(root);

        return height != -1;
    }

    private int dfs(TreeNode root) {
        if (root == null) {
            return 0;
        }

        int left = dfs(root.left);
        int right = dfs(root.right);

        if (left == -1 || right == -1 || Math.abs(left - right) > 1) {
            return -1;
        }

        return Math.max(left, right) + 1;
    }
}
